Fix availability guards and compiler warnings
# Motivation
We missed a few availability guards in our tests which caused errors when compiling for older Darwin platforms. Additionally the latest NIO release deprecated a few APIs on `NIOAsyncChannel`.

# Modification
This PR adds the missing guards and fixes all of the deprecation warnings.

# Result
We should build on all supported platforms again