Add `NWParameters` configurator to bootstraps (#230)
Allow users to provide a closure taking an `NWParameters` to customise
them before they're used to create `NWConnection`s.

### Motivation:

`NWParameters` are currently created using the provided TLS and UDP
options, and then passed over to new `NWConnection`s. However, there are
more ways in which `NWParameters` can be customised, so this new API
provides a way for users to do this.

### Modifications:

Introduce new `configureNWParameters` methods to the existing bootstraps
to allow configuring a closure for customising `NWParameters`.

### Result:

Users can now customise the `NWParameters` used to create new
`NWConnection`s.