Since iOS 18, we're occasionally transitioned to the .setup state (#242)
Since iOS 18, we're occasionally transitioned to the .setup state

### Motivation:

Closes #237.

We don't know how to reproduce this problem though, only that it occurs
extremely rarely on iOS 18 & later (including 26).

### Modifications:

Simply removed a precondition failure if we're informed of a transition
to the `.setup` state. I see no harm in this — it wasn't previously
happening, and if it happens now then we'll do nothing.

The only concern might be if this represents some kind of "reset", after
which we'll then pass back through states we've already visited — that
could conceivably break us or downstream consumers. I don't know how to
know that, test for it, or otherwise deal with it, though…

### Result:

Hopefully, we just won't crash any more.